% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHospitals.R
\docType{data}
\name{AllHospitals}
\alias{AllHospitals}
\title{Data from all German Hospitals over the years 2015 until 2019.}
\format{
A \code{tibble} with 11326 rows and 40 columns:
\describe{
  \item{\code{year}}{}
}
}
\source{

}
\usage{
AllHospitals
}
\description{
This dataset contains ...
}
\keyword{datasets}
