% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbgbaApp.R
\docType{package}
\name{qbgbaApp}
\alias{qbgbaApp}
\title{qbgbaApp: A golem package from the "Klinik-DB" project, which provides
example analyses for the German "Qualitaetsberichte der Krankenhaeuser"
obtained from the "Gemeinsamer Bundesausschuss" (GBA).}
\description{
This package aims at ...
}
